move test grpc server to seperate package
