+++ internal/impl/grpc/client_output_test.go
@@ -3,303 +3,25 @@ package grpc
 import (
 	"bytes"
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"errors"
 	"fmt"
-	"io"
 	"log/slog"
-	"net"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"strings"
-	"sync"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/warpstreamlabs/bento/internal/component/testutil"
-	test_server "github.com/warpstreamlabs/bento/internal/impl/grpc/grpc_test_server"
 	"github.com/warpstreamlabs/bento/internal/manager/mock"
 	"github.com/warpstreamlabs/bento/internal/message"
 	"github.com/warpstreamlabs/bento/internal/transaction"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/health"
-	healthgrpc "google.golang.org/grpc/health/grpc_health_v1"
-	"google.golang.org/grpc/metadata"
-	"google.golang.org/grpc/reflection"
-	"google.golang.org/grpc/status"
 
 	_ "github.com/warpstreamlabs/bento/public/components/io"
 	_ "github.com/warpstreamlabs/bento/public/components/pure"
 	"github.com/warpstreamlabs/bento/public/service"
+
+	test_server "github.com/warpstreamlabs/bento/internal/impl/grpc/grpc_test_server/test_server"
 )
-
-//------------------------------------------------------------------------------
-
-type testServer struct {
-	test_server.UnimplementedGreeterServer
-
-	reflection   bool
-	tls          bool
-	oauth2       bool
-	healthCheck  bool
-	returnErrors bool
-
-	oauthAddress string
-
-	mu                           sync.Mutex
-	SayHelloInvocations          int
-	SayMultiHellosInvocations    int
-	SayHelloHowAreYouInvocations int
-	SayHelloBidiInvocations      int
-	receivedMetadata             []metadata.MD
-	port                         int
-}
-
-func startGRPCServer(t *testing.T, opts ...testServerOpt) *testServer {
-	t.Helper()
-
-	testServer := &testServer{}
-
-	for _, o := range opts {
-		o(testServer)
-	}
-
-	lis, err := net.Listen("tcp", "localhost:0")
-	assert.NoError(t, err)
-
-	testServer.port = lis.Addr().(*net.TCPAddr).Port
-
-	serverOpts := []grpc.ServerOption{}
-
-	if testServer.tls {
-		serverCert, err := tls.LoadX509KeyPair(
-			"./grpc_test_server/certs/server.pem",
-			"./grpc_test_server/certs/server.key",
-		)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		caCert, err := os.ReadFile("./grpc_test_server/certs/ca.pem")
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		certPool := x509.NewCertPool()
-		if !certPool.AppendCertsFromPEM(caCert) {
-			t.Fatal("")
-		}
-
-		tlsConfig := &tls.Config{
-			Certificates: []tls.Certificate{serverCert},
-			ClientAuth:   tls.RequireAndVerifyClientCert,
-			ClientCAs:    certPool,
-		}
-
-		creds := credentials.NewTLS(tlsConfig)
-		serverOpts = append(serverOpts, grpc.Creds(creds))
-	}
-
-	if testServer.oauth2 {
-		tsOAuth2 := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-			username, password, ok := r.BasicAuth()
-			require.True(t, ok)
-			require.Equal(t, "fookey", username)
-			require.Equal(t, "foosecret", password)
-
-			b, err := io.ReadAll(r.Body)
-			require.NoError(t, err)
-			assert.Equal(t, "grant_type=client_credentials", string(b))
-
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			resp := `{"access_token":"some-secret-token","token_type":"Bearer","expires_in":3600}`
-			_, _ = w.Write([]byte(resp))
-		}))
-
-		testServer.oauthAddress = tsOAuth2.URL
-		serverOpts = append(serverOpts, grpc.UnaryInterceptor(ensureValidToken))
-	}
-
-	s := grpc.NewServer(serverOpts...)
-
-	if testServer.healthCheck {
-		hc := health.NewServer()
-		healthgrpc.RegisterHealthServer(s, hc)
-		hc.SetServingStatus("", healthgrpc.HealthCheckResponse_SERVING)
-	}
-
-	if testServer.reflection {
-		reflection.Register(s)
-	}
-
-	test_server.RegisterGreeterServer(s, testServer)
-	go func() {
-		err := s.Serve(lis)
-		require.NoError(t, err)
-	}()
-	return testServer
-}
-
-//------------------------------------------------------------------------------
-
-func (s *testServer) SayHello(ctx context.Context, in *test_server.HelloRequest) (*test_server.HelloReply, error) {
-	s.mu.Lock()
-	s.SayHelloInvocations++
-	if md, ok := metadata.FromIncomingContext(ctx); ok {
-		s.receivedMetadata = append(s.receivedMetadata, md)
-	}
-	s.mu.Unlock()
-	if s.returnErrors {
-		return nil, errors.New("ERROR :( ")
-	}
-	return &test_server.HelloReply{Message: "Hello " + in.GetName()}, nil
-}
-
-func (s *testServer) SayMultipleHellos(stream test_server.Greeter_SayMultipleHellosServer) error {
-	s.mu.Lock()
-	s.SayMultiHellosInvocations++
-	if md, ok := metadata.FromIncomingContext(stream.Context()); ok {
-		s.receivedMetadata = append(s.receivedMetadata, md)
-	}
-	s.mu.Unlock()
-	names := []string{}
-
-	for {
-		in, err := stream.Recv()
-		if err == io.EOF {
-			if s.returnErrors {
-				return errors.New("ERROR :( ")
-			}
-
-			return stream.SendAndClose(&test_server.HelloReply{
-				Message: "Hello " + strings.Join(names, ", "),
-			})
-		}
-		if err != nil {
-			return err
-		}
-		names = append(names, in.Name)
-	}
-}
-
-func (s *testServer) SayHelloHowAreYou(in *test_server.HelloRequest, stream test_server.Greeter_SayHelloHowAreYouServer) error {
-	s.mu.Lock()
-	s.SayHelloHowAreYouInvocations++
-	if md, ok := metadata.FromIncomingContext(stream.Context()); ok {
-		s.receivedMetadata = append(s.receivedMetadata, md)
-	}
-	s.mu.Unlock()
-
-	if s.returnErrors {
-		return errors.New("ERROR :( ")
-	}
-
-	helloMsg := &test_server.HelloReply{Message: "Hello " + in.GetName()}
-	err := stream.Send(helloMsg)
-	if err != nil {
-		return err
-	}
-	howAreYouMsg := &test_server.HelloReply{Message: "How are you, " + in.GetName() + "?"}
-	err = stream.Send(howAreYouMsg)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (s *testServer) SayHelloBidi(stream grpc.BidiStreamingServer[test_server.HelloRequest, test_server.HelloReply]) error {
-	s.mu.Lock()
-	s.SayHelloBidiInvocations++
-	if md, ok := metadata.FromIncomingContext(stream.Context()); ok {
-		s.receivedMetadata = append(s.receivedMetadata, md)
-	}
-	s.mu.Unlock()
-
-	if s.returnErrors {
-		return errors.New("ERROR :( ")
-	}
-
-	return nil
-}
-
-//------------------------------------------------------------------------------
-
-var (
-	errMissingMetadata = status.Errorf(codes.InvalidArgument, "missing metadata")
-	errInvalidToken    = status.Errorf(codes.Unauthenticated, "invalid token")
-)
-
-func valid(authorization []string) bool {
-	if len(authorization) < 1 {
-		return false
-	}
-	token := strings.TrimPrefix(authorization[0], "Bearer ")
-
-	return token == "some-secret-token"
-}
-
-func ensureValidToken(ctx context.Context, req any, _ *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (any, error) {
-	md, ok := metadata.FromIncomingContext(ctx)
-	if !ok {
-		return nil, errMissingMetadata
-	}
-
-	if !valid(md["authorization"]) {
-		return nil, errInvalidToken
-	}
-
-	return handler(ctx, req)
-}
-
-//------------------------------------------------------------------------------
-
-type testServerOpt func(*testServer)
-
-func withReflection() testServerOpt {
-	return func(ts *testServer) {
-		ts.reflection = true
-	}
-}
-
-func withTLS() testServerOpt {
-	return func(ts *testServer) {
-		ts.tls = true
-	}
-}
-
-func withHealthCheck() testServerOpt {
-	return func(ts *testServer) {
-		ts.healthCheck = true
-	}
-}
-
-func withOAuth2() testServerOpt {
-	return func(ts *testServer) {
-		ts.oauth2 = true
-	}
-}
-
-func withReturnErrors() testServerOpt {
-	return func(ts *testServer) {
-		ts.returnErrors = true
-	}
-}
-
-func (s *testServer) getReceivedMetadata() []metadata.MD {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	cp := make([]metadata.MD, len(s.receivedMetadata))
-	copy(cp, s.receivedMetadata)
-	return cp
-}
 
 //------------------------------------------------------------------------------
 
@@ -308,14 +30,14 @@ var names = []string{"Alice", "Bob", "Carol", "Dan"}
 
 func TestGrpcClientWriter(t *testing.T) {
 	tests := map[string]struct {
-		grpcServerOpts   []testServerOpt
+		grpcServerOpts   []test_server.TestServerOpt
 		confFormatString string
-		formatArgs       func(*testServer) []any
-		invocations      func(*testServer) int
+		formatArgs       func(*test_server.TestServer) []any
+		invocations      func(*test_server.TestServer) int
 		expInvocations   int
 	}{
 		"Unary Reflection": {
-			grpcServerOpts: []testServerOpt{withReflection()},
+			grpcServerOpts: []test_server.TestServerOpt{test_server.WithReflection()},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
@@ -323,38 +45,38 @@ grpc_client:
   method: SayHello
   reflection: true
 `,
-			formatArgs: func(ts *testServer) []any {
-				return []any{ts.port}
+			formatArgs: func(ts *test_server.TestServer) []any {
+				return []any{ts.Port}
 			},
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayHelloInvocations
 			},
 			expInvocations: 4,
 		},
 		"Unary Proto File": {
-			grpcServerOpts: []testServerOpt{},
+			grpcServerOpts: []test_server.TestServerOpt{},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
   service: helloworld.Greeter
   method: SayHello
   proto_files: 
-    - "./grpc_test_server/helloworld.proto"
+    - "./grpc_test_server/helloworld/helloworld.proto"
 `,
-			formatArgs: func(ts *testServer) []any {
-				return []any{ts.port}
+			formatArgs: func(ts *test_server.TestServer) []any {
+				return []any{ts.Port}
 			},
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayHelloInvocations
 			},
 			expInvocations: 4,
 		},
 		"Client Stream Reflection": {
-			grpcServerOpts: []testServerOpt{withReflection()},
+			grpcServerOpts: []test_server.TestServerOpt{test_server.WithReflection()},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
@@ -365,18 +87,18 @@ grpc_client:
   batching:
     count: 4
 `,
-			formatArgs: func(ts *testServer) []any {
-				return []any{ts.port}
+			formatArgs: func(ts *test_server.TestServer) []any {
+				return []any{ts.Port}
 			},
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayMultiHellosInvocations
 			},
 			expInvocations: 1,
 		},
 		"Server Stream Reflection": {
-			grpcServerOpts: []testServerOpt{withReflection()},
+			grpcServerOpts: []test_server.TestServerOpt{test_server.WithReflection()},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
@@ -385,18 +107,18 @@ grpc_client:
   reflection: true
   rpc_type: server_stream
 `,
-			formatArgs: func(ts *testServer) []any {
-				return []any{ts.port}
+			formatArgs: func(ts *test_server.TestServer) []any {
+				return []any{ts.Port}
 			},
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayHelloHowAreYouInvocations
 			},
 			expInvocations: 4,
 		},
 		"Bidirectional Reflection": {
-			grpcServerOpts: []testServerOpt{withReflection()},
+			grpcServerOpts: []test_server.TestServerOpt{test_server.WithReflection()},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
@@ -407,18 +129,18 @@ grpc_client:
   batching:
     count: 4
 `,
-			formatArgs: func(ts *testServer) []any {
-				return []any{ts.port}
+			formatArgs: func(ts *test_server.TestServer) []any {
+				return []any{ts.Port}
 			},
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayHelloBidiInvocations
 			},
 			expInvocations: 1,
 		},
 		"TLS": {
-			grpcServerOpts: []testServerOpt{withReflection(), withTLS()},
+			grpcServerOpts: []test_server.TestServerOpt{test_server.WithReflection(), test_server.WithTLS()},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
@@ -427,24 +149,24 @@ grpc_client:
   reflection: true
   tls:
     enabled: true
-    root_cas_file: ./grpc_test_server/certs/ca.pem
+    root_cas_file: ./grpc_test_server/helloworld/certs/ca.pem
     client_certs:
-      - cert_file: ./grpc_test_server/certs/client.pem
-        key_file: ./grpc_test_server/certs/client.key
+      - cert_file: ./grpc_test_server/helloworld/certs/client.pem
+        key_file: ./grpc_test_server/helloworld/certs/client.key
     skip_cert_verify: false
 `,
-			formatArgs: func(ts *testServer) []any {
-				return []any{ts.port}
+			formatArgs: func(ts *test_server.TestServer) []any {
+				return []any{ts.Port}
 			},
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayHelloInvocations
 			},
 			expInvocations: 4,
 		},
 		"TLS and oAuth2": {
-			grpcServerOpts: []testServerOpt{withReflection(), withTLS(), withOAuth2()},
+			grpcServerOpts: []test_server.TestServerOpt{test_server.WithReflection(), test_server.WithTLS(), test_server.WithOAuth2()},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
@@ -458,18 +180,18 @@ grpc_client:
     client_secret: foosecret
   tls:
     enabled: true
-    root_cas_file: ./grpc_test_server/certs/ca.pem
+    root_cas_file: ./grpc_test_server/helloworld/certs/ca.pem
     client_certs:
-      - cert_file: ./grpc_test_server/certs/client.pem
-        key_file: ./grpc_test_server/certs/client.key
+      - cert_file: ./grpc_test_server/helloworld/certs/client.pem
+        key_file: ./grpc_test_server/helloworld/certs/client.key
     skip_cert_verify: false
 `,
-			formatArgs: func(ts *testServer) []any {
-				return []any{ts.port, ts.oauthAddress}
+			formatArgs: func(ts *test_server.TestServer) []any {
+				return []any{ts.Port, ts.OAuthAddress}
 			},
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayHelloInvocations
 			},
 			expInvocations: 4,
@@ -480,7 +202,7 @@ grpc_client:
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			testServer := startGRPCServer(t, test.grpcServerOpts...)
+			testServer := test_server.StartGRPCServer(t, test.grpcServerOpts...)
 
 			yamlConf := fmt.Sprintf(test.confFormatString, test.formatArgs(testServer)...)
 
@@ -514,15 +236,15 @@ grpc_client:
 
 func TestGrpcClientWriterSyncResponse(t *testing.T) {
 	tests := map[string]struct {
-		grpcServerOpts   []testServerOpt
+		grpcServerOpts   []test_server.TestServerOpt
 		confFormatString string
-		invocations      func(*testServer) int
+		invocations      func(*test_server.TestServer) int
 		expInvocations   int
 		inputs           []string
 		waitForBatch     bool
 	}{
 		"Unary Sync Response": {
-			grpcServerOpts: []testServerOpt{withReflection()},
+			grpcServerOpts: []test_server.TestServerOpt{test_server.WithReflection()},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
@@ -531,16 +253,16 @@ grpc_client:
   propagate_response: true
   reflection: true
 `,
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayHelloInvocations
 			},
 			expInvocations: 4,
 			inputs:         namesInputTestData,
 		},
 		"Unary Sync Response Protofile": {
-			grpcServerOpts: []testServerOpt{},
+			grpcServerOpts: []test_server.TestServerOpt{},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
@@ -548,11 +270,11 @@ grpc_client:
   method: SayHello
   propagate_response: true
   proto_files:
-    - "./grpc_test_server/helloworld.proto"
+    - "./grpc_test_server/helloworld/helloworld.proto"
 `,
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayHelloInvocations
 			},
 			expInvocations: 4,
@@ -564,9 +286,9 @@ grpc_client:
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			testServer := startGRPCServer(t, test.grpcServerOpts...)
+			testServer := test_server.StartGRPCServer(t, test.grpcServerOpts...)
 
-			yamlConf := fmt.Sprintf(test.confFormatString, testServer.port)
+			yamlConf := fmt.Sprintf(test.confFormatString, testServer.Port)
 
 			sendChan, receiveChan, err := startGrpcClientOutput(t, yamlConf)
 			require.NoError(t, err)
@@ -601,7 +323,7 @@ grpc_client:
 }
 
 func TestGrpcClientWriterClientStreamSyncResponse(t *testing.T) {
-	testServer := startGRPCServer(t, withReflection())
+	testServer := test_server.StartGRPCServer(t, test_server.WithReflection())
 
 	yamlConf := fmt.Sprintf(`
 grpc_client:
@@ -613,7 +335,7 @@ grpc_client:
   rpc_type: client_stream
   batching:
     count: 4
-`, testServer.port)
+`, testServer.Port)
 
 	sendChan, receiveChan, err := startGrpcClientOutput(t, yamlConf)
 	require.NoError(t, err)
@@ -653,7 +375,7 @@ grpc_client:
 }
 
 func TestGrpcClientWriterServerStreamSyncResponse(t *testing.T) {
-	testServer := startGRPCServer(t, withReflection())
+	testServer := test_server.StartGRPCServer(t, test_server.WithReflection())
 
 	yamlConf := fmt.Sprintf(`
 grpc_client:
@@ -663,7 +385,7 @@ grpc_client:
   reflection: true
   rpc_type: server_stream
   propagate_response: true
-`, testServer.port)
+`, testServer.Port)
 
 	sendChan, receiveChan, err := startGrpcClientOutput(t, yamlConf)
 	require.NoError(t, err)
@@ -699,7 +421,7 @@ grpc_client:
 }
 
 func TestGrpcClientWriterHealthCheck(t *testing.T) {
-	testServer := startGRPCServer(t, withReflection(), withHealthCheck())
+	testServer := test_server.StartGRPCServer(t, test_server.WithReflection(), test_server.WithHealthCheck())
 
 	yamlConf := fmt.Sprintf(`
 address: localhost:%v
@@ -709,7 +431,7 @@ reflection: true
 health_check:
   enabled: true
   service: ""
-`, testServer.port)
+`, testServer.Port)
 
 	pConf, err := grcpClientOutputSpec().ParseYAML(yamlConf, nil)
 	require.NoError(t, err)
@@ -723,7 +445,7 @@ health_check:
 }
 
 func TestGrpcClientWriterUnableToFindMethodErr(t *testing.T) {
-	testServer := startGRPCServer(t, withReflection())
+	testServer := test_server.StartGRPCServer(t, test_server.WithReflection())
 
 	sb := service.NewStreamBuilder()
 
@@ -739,7 +461,7 @@ output:
     service: helloworld.Greeter
     method: DoesNotExist
     reflection: true
-`, testServer.port))
+`, testServer.Port))
 	require.NoError(t, err)
 
 	buffer := new(bytes.Buffer)
@@ -772,7 +494,7 @@ output:
 }
 
 func TestGrpcClientWriterBrokenProtoFile(t *testing.T) {
-	testServer := startGRPCServer(t, withReflection())
+	testServer := test_server.StartGRPCServer(t, test_server.WithReflection())
 
 	sb := service.NewStreamBuilder()
 
@@ -788,8 +510,8 @@ output:
     service: helloworld.Greeter
     method: SayHello
     proto_files: 
-      - "./grpc_test_server/fail_parse.proto"
-`, testServer.port))
+      - "./grpc_test_server/helloworld/fail_parse.proto"
+`, testServer.Port))
 	require.NoError(t, err)
 
 	buffer := new(bytes.Buffer)
@@ -896,13 +618,13 @@ grpc_client:
 
 func TestGrpcClientWriterBatchErrors(t *testing.T) {
 	tests := map[string]struct {
-		grpcServerOpts   []testServerOpt
+		grpcServerOpts   []test_server.TestServerOpt
 		confFormatString string
-		formatArgs       func(*testServer) []any
-		invocations      func(*testServer) int
+		formatArgs       func(*test_server.TestServer) []any
+		invocations      func(*test_server.TestServer) int
 	}{
 		"Unary Reflection Errors": {
-			grpcServerOpts: []testServerOpt{withReflection(), withReturnErrors()},
+			grpcServerOpts: []test_server.TestServerOpt{test_server.WithReflection(), test_server.WithReturnErrors()},
 			confFormatString: `
 grpc_client:
   address: localhost:%v
@@ -910,12 +632,12 @@ grpc_client:
   method: SayHello
   reflection: true
 `,
-			formatArgs: func(ts *testServer) []any {
-				return []any{ts.port}
+			formatArgs: func(ts *test_server.TestServer) []any {
+				return []any{ts.Port}
 			},
-			invocations: func(ts *testServer) int {
-				ts.mu.Lock()
-				defer ts.mu.Unlock()
+			invocations: func(ts *test_server.TestServer) int {
+				ts.Mu.Lock()
+				defer ts.Mu.Unlock()
 				return ts.SayHelloInvocations
 			},
 		},
@@ -925,7 +647,7 @@ grpc_client:
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			testServer := startGRPCServer(t, test.grpcServerOpts...)
+			testServer := test_server.StartGRPCServer(t, test.grpcServerOpts...)
 
 			yamlConf := fmt.Sprintf(test.confFormatString, test.formatArgs(testServer)...)
 
@@ -954,7 +676,7 @@ grpc_client:
 }
 
 func TestGrpcClientWriterMetadataUnary(t *testing.T) {
-	ts := startGRPCServer(t, withReflection())
+	ts := test_server.StartGRPCServer(t, test_server.WithReflection())
 
 	yamlConf := fmt.Sprintf(`
 grpc_client:
@@ -965,7 +687,7 @@ grpc_client:
   metadata:
     x-custom-key: custom-value
     x-request-name: fixed-name
-`, ts.port)
+`, ts.Port)
 
 	sendChan, receiveChan, err := startGrpcClientOutput(t, yamlConf)
 	require.NoError(t, err)
@@ -992,7 +714,7 @@ grpc_client:
 		return ts.SayHelloInvocations == len(namesInputTestData)
 	}, time.Second*10, time.Millisecond*20)
 
-	received := ts.getReceivedMetadata()
+	received := ts.GetReceivedMetadata()
 	require.Len(t, received, len(namesInputTestData))
 	for _, md := range received {
 		assert.Equal(t, []string{"custom-value"}, md.Get("x-custom-key"))
@@ -1001,7 +723,7 @@ grpc_client:
 }
 
 func TestGrpcClientWriterMetadataClientStream(t *testing.T) {
-	ts := startGRPCServer(t, withReflection())
+	ts := test_server.StartGRPCServer(t, test_server.WithReflection())
 
 	yamlConf := fmt.Sprintf(`
 grpc_client:
@@ -1014,7 +736,7 @@ grpc_client:
     x-stream-id: my-stream
   batching:
     count: 4
-`, ts.port)
+`, ts.Port)
 
 	sendChan, receiveChan, err := startGrpcClientOutput(t, yamlConf)
 	require.NoError(t, err)
@@ -1038,18 +760,18 @@ grpc_client:
 	}
 
 	assert.Eventually(t, func() bool {
-		ts.mu.Lock()
-		defer ts.mu.Unlock()
+		ts.Mu.Lock()
+		defer ts.Mu.Unlock()
 		return ts.SayMultiHellosInvocations == 1
 	}, time.Second*10, time.Millisecond*20)
 
-	received := ts.getReceivedMetadata()
+	received := ts.GetReceivedMetadata()
 	require.Len(t, received, 1)
 	assert.Equal(t, []string{"my-stream"}, received[0].Get("x-stream-id"))
 }
 
 func TestGrpcClientWriterMetadataServerStream(t *testing.T) {
-	ts := startGRPCServer(t, withReflection())
+	ts := test_server.StartGRPCServer(t, test_server.WithReflection())
 
 	yamlConf := fmt.Sprintf(`
 grpc_client:
@@ -1060,7 +782,7 @@ grpc_client:
   rpc_type: server_stream
   metadata:
     x-server-stream: "true"
-`, ts.port)
+`, ts.Port)
 
 	sendChan, receiveChan, err := startGrpcClientOutput(t, yamlConf)
 	require.NoError(t, err)
@@ -1084,12 +806,12 @@ grpc_client:
 	}
 
 	assert.Eventually(t, func() bool {
-		ts.mu.Lock()
-		defer ts.mu.Unlock()
+		ts.Mu.Lock()
+		defer ts.Mu.Unlock()
 		return ts.SayHelloHowAreYouInvocations == len(namesInputTestData)
 	}, time.Second*10, time.Millisecond*20)
 
-	received := ts.getReceivedMetadata()
+	received := ts.GetReceivedMetadata()
 	require.Len(t, received, len(namesInputTestData))
 	for _, md := range received {
 		assert.Equal(t, []string{"true"}, md.Get("x-server-stream"))
@@ -1097,7 +819,7 @@ grpc_client:
 }
 
 func TestGrpcClientWriterMetadataBidi(t *testing.T) {
-	ts := startGRPCServer(t, withReflection())
+	ts := test_server.StartGRPCServer(t, test_server.WithReflection())
 
 	yamlConf := fmt.Sprintf(`
 grpc_client:
@@ -1110,7 +832,7 @@ grpc_client:
     x-bidi-key: bidi-value
   batching:
     count: 4
-`, ts.port)
+`, ts.Port)
 
 	sendChan, receiveChan, err := startGrpcClientOutput(t, yamlConf)
 	require.NoError(t, err)
@@ -1134,18 +856,18 @@ grpc_client:
 	}
 
 	assert.Eventually(t, func() bool {
-		ts.mu.Lock()
-		defer ts.mu.Unlock()
+		ts.Mu.Lock()
+		defer ts.Mu.Unlock()
 		return ts.SayHelloBidiInvocations == 1
 	}, time.Second*10, time.Millisecond*20)
 
-	received := ts.getReceivedMetadata()
+	received := ts.GetReceivedMetadata()
 	require.Len(t, received, 1)
 	assert.Equal(t, []string{"bidi-value"}, received[0].Get("x-bidi-key"))
 }
 
 func TestGrpcClientWriterMetadataInterpolation(t *testing.T) {
-	ts := startGRPCServer(t, withReflection())
+	ts := test_server.StartGRPCServer(t, test_server.WithReflection())
 
 	yamlConf := fmt.Sprintf(`
 grpc_client:
@@ -1155,7 +877,7 @@ grpc_client:
   reflection: true
   metadata:
     x-name: ${! json("name") }
-`, ts.port)
+`, ts.Port)
 
 	sendChan, receiveChan, err := startGrpcClientOutput(t, yamlConf)
 	require.NoError(t, err)
@@ -1182,7 +904,7 @@ grpc_client:
 		return ts.SayHelloInvocations == len(namesInputTestData)
 	}, time.Second*10, time.Millisecond*20)
 
-	received := ts.getReceivedMetadata()
+	received := ts.GetReceivedMetadata()
 	require.Len(t, received, len(namesInputTestData))
 
 	gotNames := make([]string, 0, len(received))
