+++ internal/impl/grpc/grpc_test_server/helloworld/helloworld.proto
